(closes #27) Fix bug in handling capitalized nicks
This fixes a bug where nicks that are capitalized are not removed from
the map of nicks to user structs in the global state. This leads to
strange behavior where only the original user of a capitalized nick
may change to that nick, and that WHO queries will return entries when
queried against old nicks.